refactor(ui): rename StudyTreeView to VariantsTree
