+++ webapp/src/routes/_authenticated/studies/$studyId/-components/VariantsTree/constants.ts
