+++ webapp/src/routes/_authenticated/studies/$studyId/-components/VariantsTree/utils.ts
